server/requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
djangorestframework==3.4.6
django-extensions==1.7.4
django-cors-headers==1.1.0
